Removing pxtlib dependency on typescript
