+++ Jakefile.js
@@ -63,6 +63,7 @@ task('testpkgconflicts', ['built/pxt.js'], { async: true }, function () {
 ju.catFiles('built/pxt.js', [
     "node_modules/typescript/lib/typescript.js",
     "built/pxtlib.js",
+    "built/pxttypescript.js",
     "built/pxtsim.js",
     "built/cli.js"
 ],
@@ -91,14 +92,15 @@ file('built/pxt-common.json', expand(['libs/pxt-common'], ".ts"), function () {
     fs.writeFileSync(this.name, JSON.stringify(std, null, 4))
 })
 
-compileDir("pxtlib", ["built/typescriptServices.d.ts"])
+compileDir("pxtlib", "built/typescriptServices.d.ts")
+compileDir("pxttypescript", ["built/pxtlib.js"])
 compileDir("pxtwinrt", ["built/pxtlib.js"])
 compileDir("pxtblocks", ["built/pxtlib.js"])
 compileDir("pxtrunner", ["built/pxtlib.js", "built/pxtsim.js", "built/pxtblocks.js"])
 compileDir("pxtsim", ["built/pxtlib.js", "built/pxtblocks.js"])
 compileDir("pxteditor", ["built/pxtlib.js", "built/pxtblocks.js"])
 compileDir("cli", ["built/pxtlib.js", "built/pxtsim.js"])
-compileDir("backendutils", ['pxtlib/emitter/util.ts', 'pxtlib/docsrender.ts'])
+compileDir("backendutils", ['pxtlib/util.ts', 'pxtlib/docsrender.ts'])
 
 task("travis", ["lint", "test", "upload"])
 
@@ -123,7 +125,7 @@ task("lint", [], { async: true }, function () {
         "pxtblocks",
         "pxteditor",
         "pxtlib",
-        "pxtlib/emitter",
+        "pxttypescript/emitter",
         "pxtrunner",
         "pxtsim",
         "pxtwinrt",
@@ -218,6 +220,7 @@ file('built/localization.json', ju.expand1(
 
 task('wapp', [
     "built/web/pxtlib.js",
+    "built/web/pxttypescript.js",
     "built/web/pxtsim.js",
     "built/web/pxtblocks.js",
     "built/web/pxteditor.js",
@@ -232,6 +235,7 @@ task('wapp', [
 
 file("built/web/pxtlib.js", [
     "built/pxtlib.js",
+    "built/pxttypescript.js",
     "built/pxtblocks.js",
     "built/pxtsim.js",
     "built/pxtrunner.js",
@@ -244,6 +248,7 @@ file("built/web/pxtlib.js", [
     jake.cpR("node_modules/fuse.js/src/fuse.min.js", "built/web/fuse.min.js")
 
     jake.cpR("built/pxtlib.js", "built/web/")
+    jake.cpR("built/pxttypescript.js", "built/web/")
     jake.cpR("built/pxtblocks.js", "built/web/")
     jake.cpR("built/pxtsim.js", "built/web/")
     jake.cpR("built/pxtrunner.js", "built/web/")
