+++ pxtlib/cmds.ts
@@ -1,5 +1,5 @@
 namespace pxt.commands {
     // overriden by targets
-    export let deployCoreAsync: (r: pxtc.CompileResult) => Promise<void> = undefined;
+    export let deployCoreAsync: (r: ts.pxtc.CompileResult) => Promise<void> = undefined;
     export let browserDownloadAsync: (text: string, name: string, contentType: string) => Promise<void> = undefined;
 }
