+++ pxtlib/docsrender.ts
@@ -1,7 +1,7 @@
 /// <reference path='../typings/globals/marked/index.d.ts' />
 /// <reference path='../typings/globals/highlightjs/index.d.ts' />
 /// <reference path='../localtypings/pxtarget.d.ts' />
-/// <reference path="emitter/util.ts"/>
+/// <reference path="util.ts"/>
 
 namespace pxt.docs {
     declare var require: any;
