+++ pxtlib/util.ts
@@ -1,4 +1,4 @@
-/// <reference path="../../typings/globals/bluebird/index.d.ts"/>
+/// <reference path="../typings/globals/bluebird/index.d.ts"/>
 
 namespace ts.pxtc {
     export var __dummy = 42;
