+++ pxtsim/state/toggle.ts
@@ -1,0 +1,5 @@
+namespace pxsim {
+    export class ToggleState {
+        on = false;
+    }
+}
