+++ pxttypescript/emitter/avr.ts
@@ -1,4 +1,3 @@
-/// <reference path="assembler.ts"/>
 
 /* Docs:
     *
