+++ pxttypescript/emitter/backthumb.ts
