+++ pxttypescript/emitter/emitter.ts
@@ -2,18 +2,6 @@
 /// <reference path="../../localtypings/pxtpackage.d.ts"/>
 
 namespace ts.pxtc {
-    export const assert = Util.assert;
-    export const oops = Util.oops;
-    export import U = pxtc.Util;
-
-    export const ON_START_TYPE = "pxt-on-start";
-    export const ON_START_COMMENT = U.lf("on start");
-    export const TS_STATEMENT_TYPE = "typescript_statement";
-    export const TS_OUTPUT_TYPE = "typescript_expression";
-    export const BINARY_JS = "binary.js";
-    export const BINARY_HEX = "binary.hex";
-    export const BINARY_ASM = "binary.asm";
-    export const BINARY_UF2 = "binary.uf2";
 
     let EK = ir.EK;
     export const SK = SyntaxKind;
@@ -269,75 +257,6 @@ namespace ts.pxtc {
         Int32,
     }
 
-    export interface CommentAttrs {
-        debug?: boolean; // requires ?dbg=1
-        shim?: string;
-        enumval?: string;
-        helper?: string;
-        help?: string;
-        async?: boolean;
-        promise?: boolean;
-        hidden?: boolean;
-        callingConvention: ir.CallingConvention;
-        block?: string; // format of the block
-        blockId?: string; // unique id of the block
-        blockGap?: string; // pixels in toolbox after the block is inserted
-        blockExternalInputs?: boolean; // force external inputs
-        blockImportId?: string;
-        blockBuiltin?: boolean;
-        blockNamespace?: string;
-        blockIdentity?: string;
-        blockAllowMultiple?: boolean; // override single block behavior for events
-        blockHidden?: boolean; // not available directly in toolbox
-        blockImage?: boolean; // for enum variable, specifies that it should use an image from a predefined location
-        blockFieldEditor?: string; // Custom field editor
-        fixedInstances?: boolean;
-        fixedInstance?: boolean;
-        indexedInstanceNS?: string;
-        indexedInstanceShim?: string;
-        defaultInstance?: string;
-        autoCreate?: string;
-        noRefCounting?: boolean;
-        color?: string;
-        icon?: string;
-        imageLiteral?: number;
-        weight?: number;
-        parts?: string;
-        trackArgs?: number[];
-        advanced?: boolean;
-        deprecated?: boolean;
-        useEnumVal?: boolean; // for conversion from typescript to blocks with enumVal
-        // On block
-        subcategory?: string;
-        // On namepspace
-        subcategories?: string[];
-
-        // on interfaces
-        indexerGet?: string;
-        indexerSet?: string;
-
-        mutate?: string;
-        mutateText?: string;
-        mutatePrefix?: string;
-        mutateDefaults?: string;
-        mutatePropertyEnum?: string;
-
-        _name?: string;
-        _source?: string;
-        jsDoc?: string;
-        paramHelp?: pxt.Map<string>;
-        // foo.defl=12 -> paramDefl: { foo: "12" }
-        paramDefl: pxt.Map<string>;
-
-        paramMin?: pxt.Map<string>; // min range
-        paramMax?: pxt.Map<string>; // max range
-        // Map for custom field editor parameters
-        blockFieldEditorParams?: pxt.Map<string>;
-    }
-
-    const numberAttributes = ["weight", "imageLiteral"]
-    const booleanAttributes = ["advanced"]
-
     export interface CallInfo {
         decl: Declaration;
         qName: string;
@@ -398,82 +317,6 @@ namespace ts.pxtc {
         } else {
             return cmtCore(node)
         }
-    }
-
-    export function parseCommentString(cmt: string): CommentAttrs {
-        let res: CommentAttrs = {
-            paramDefl: {},
-            callingConvention: ir.CallingConvention.Plain,
-            _source: cmt
-        }
-        let didSomething = true
-        while (didSomething) {
-            didSomething = false
-            cmt = cmt.replace(/\/\/%[ \t]*([\w\.]+)(=(("[^"\n]+")|'([^'\n]+)'|([^\s]*)))?/,
-                (f: string, n: string, d0: string, d1: string,
-                    v0: string, v1: string, v2: string) => {
-                    let v = v0 ? JSON.parse(v0) : (d0 ? (v0 || v1 || v2) : "true");
-                    if (U.endsWith(n, ".defl")) {
-                        res.paramDefl[n.slice(0, n.length - 5)] = v
-                    } else if (U.endsWith(n, ".min")) {
-                        if (!res.paramMin) res.paramMin = {}
-                        res.paramMin[n.slice(0, n.length - 4)] = v
-                    } else if (U.endsWith(n, ".max")) {
-                        if (!res.paramMax) res.paramMax = {}
-                        res.paramMax[n.slice(0, n.length - 4)] = v
-                    } else if (U.startsWith(n, "blockFieldEditorParams")) {
-                        if (!res.blockFieldEditorParams) res.blockFieldEditorParams = {}
-                        res.blockFieldEditorParams[n.slice(23, n.length)] = v
-                    } else {
-                        (<any>res)[n] = v;
-                    }
-                    didSomething = true
-                    return "//% "
-                })
-        }
-
-        for (let n of numberAttributes) {
-            if (typeof (res as any)[n] == "string")
-                (res as any)[n] = parseInt((res as any)[n])
-        }
-
-        for (let n of booleanAttributes) {
-            if (typeof (res as any)[n] == "string")
-                (res as any)[n] = (res as any)[n] == 'true' || (res as any)[n] == '1' ? true : false;
-        }
-
-        if (res.trackArgs) {
-            res.trackArgs = ((res.trackArgs as any) as string).split(/[ ,]+/).map(s => parseInt(s) || 0)
-        }
-
-        res.paramHelp = {}
-        res.jsDoc = ""
-        cmt = cmt.replace(/\/\*\*([^]*?)\*\//g, (full: string, doccmt: string) => {
-            doccmt = doccmt.replace(/\n\s*(\*\s*)?/g, "\n")
-            doccmt = doccmt.replace(/^\s*@param\s+(\w+)\s+(.*)$/mg, (full: string, name: string, desc: string) => {
-                res.paramHelp[name] = desc
-                return ""
-            })
-            res.jsDoc += doccmt
-            return ""
-        })
-
-        res.jsDoc = res.jsDoc.trim()
-
-        if (res.async)
-            res.callingConvention = ir.CallingConvention.Async
-        if (res.promise)
-            res.callingConvention = ir.CallingConvention.Promise
-        if (res.subcategories) {
-            try {
-                res.subcategories = JSON.parse(res.subcategories as any);
-            }
-            catch (e) {
-                res.subcategories = undefined;
-            }
-        }
-
-        return res
     }
 
     export function parseCommentsOnSymbol(symbol: Symbol): CommentAttrs {
@@ -3627,24 +3470,6 @@ ${lbl}: .short 0xffff
         }
     }
 
-    export function emptyExtInfo(): ExtensionInfo {
-        const pio = pxt.appTarget.compileService && !!pxt.appTarget.compileService.platformioIni;
-        const r: ExtensionInfo = {
-            functions: [],
-            generatedFiles: {},
-            extensionFiles: {},
-            sha: "",
-            compileData: "",
-            shimsDTS: "",
-            enumsDTS: "",
-            onlyPublic: true
-        }
-        if (pio) r.platformio = { dependencies: {} };
-        else r.yotta = { config: {}, dependencies: {} };
-        return r;
-    }
-
-
     export class Binary {
         procs: ir.Procedure[] = [];
         globals: ir.Cell[] = [];
@@ -3685,4 +3510,3 @@ ${lbl}: .short 0xffff
         }
     }
 }
-
