+++ pxttypescript/emitter/typescriptInternal.d.ts
