+++ pxttypescript/main.ts
@@ -1,0 +1,25 @@
+/// <reference path="../typings/globals/bluebird/index.d.ts"/>
+/// <reference path="../localtypings/pxtpackage.d.ts"/>
+/// <reference path="../localtypings/pxtparts.d.ts"/>
+/// <reference path="../localtypings/pxtarget.d.ts"/>
+
+namespace pxt {
+
+    /*
+    export class MainPackage
+        extends Package {
+        buildAsync(target: pxtc.CompileTarget) {
+            return this.getCompileOptionsAsync(target)
+                .then(opts => pxtc.compile(opts))
+        }
+
+        serviceAsync(op: string) {
+            return this.getCompileOptionsAsync()
+                .then(opts => {
+                    pxtc.service.performOperation("reset", {})
+                    pxtc.service.performOperation("setOpts", { options: opts })
+                    return pxtc.service.performOperation(op, {})
+                })
+        }
+    }*/
+}
