+++ pxttypescript/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "target": "es5",
+        "noImplicitAny": true,
+        "noImplicitReturns": true,
+        "declaration": true,
+        "out": "../built/pxttypescript.js",
+        "newLine": "LF",
+        "sourceMap": false
+    }
+}
