+++ webapp/public/index.html
@@ -40,7 +40,6 @@
     <script>if (typeof module === 'object') {window.module = module; module = undefined;}</script>
     <script type="text/javascript" src="/blb/lzma/lzma_worker-min.js"></script>
     <script type="text/javascript" src="/blb/jquery.js"></script>
-    <script type="text/javascript" src="/blb/typescript.js"></script>
     <script type="text/javascript" src="/blb/blockly/blockly_compressed.js"></script>
     <script type="text/javascript" src="/blb/blockly/blocks_compressed.js"></script>
     <script type="text/javascript" src="/blb/blockly/msg/js/en.js"></script>
